There was a issue when a call was made to the "getMetadataFor" method from a "loadClassMetadata" event that a new instance of the class was created for a class that was already called for before. So this would lead to problem that 2 metadataclass instance where made for the same class where one would have its discriminator map correctly load and the other would be empty. The empty ones would eventually overwrite the correct one which leaded to index not found errors in the discriminator map.
This should not be considered a permanent fix as its not really solving the issues it is just working around it. Really solving the issue would be to make sure that getMetadataFor will always return a correctly loaded discriminator map to external calls even when called from a "loadClassMetadata" event callback.