Redone the forms and there services.
Added a proxy form the only thing the proxy form does is forward the method calls except for the
getName. The getName will return the name configured in the constructor. This allows the use of the
same form with different aliased tags in the form factory. One use case would also be that you can
replace the form with a user made alternative with the same name.

The name of the user form is changed to profile. this ways we do not get intergrated_user_user as
form name. This also changed the name of the user type in the extension from  integrated_user to
user_profile.

Changed login controller to use the named form type instead of getting the type from a service.

The form.xml got services for all the forms in this bundle and de form.<name>.xml got service for
creating named form types used by the form factory.